+++ RecoTracker/LSTCore/src/alpaka/LSTEvent.h
@@ -40,7 +40,7 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE::lst {
   private:
     Queue& queue_;
     const float ptCut_;
-    const uint8_t clustSizeCut_;
+    const uint16_t clustSizeCut_;
 
     std::array<unsigned int, 6> n_minidoublets_by_layer_barrel_{};
     std::array<unsigned int, 5> n_minidoublets_by_layer_endcap_{};
@@ -94,7 +94,7 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE::lst {
 
   public:
     // Constructor used for CMSSW integration. Uses an external queue.
-    LSTEvent(bool verbose, const float pt_cut, const uint8_t clustsize_cut, Queue& q, const LSTESData<Device>* deviceESData)
+    LSTEvent(bool verbose, const float pt_cut, const uint16_t clustsize_cut, Queue& q, const LSTESData<Device>* deviceESData)
         : queue_(q),
           ptCut_(pt_cut),
           clustSizeCut_(clustsize_cut),
