+++ RecoTracker/LSTCore/src/alpaka/MiniDoublet.h
@@ -383,7 +383,7 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE::lst {
                                                      float zUpper,
                                                      float rtUpper,
                                                      const float ptCut,
-                                                     const uint8_t clustSizeCut) {
+                                                     const uint16_t clustSizeCut) {
     if (0 == clustSizeCut) {
       return false;
     }
@@ -516,7 +516,7 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE::lst {
                                                      float zUpper,
                                                      float rtUpper,
                                                      const float ptCut,
-                                                     const uint8_t clustSizeCut) {
+                                                     const uint16_t clustSizeCut) {
     // There are series of cuts that applies to mini-doublet in a "endcap" region
     // Cut #1 : dz cut. The dz difference can't be larger than 1cm. (max separation is 4mm for modules in the endcap)
     // Ref to original code: https://github.com/slava77/cms-tkph2-ntuple/blob/184d2325147e6930030d3d1f780136bc2dd29ce6/doubletAnalysis.C#L3093
@@ -633,7 +633,7 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE::lst {
                                                float zUpper,
                                                float rtUpper,
                                                const float ptCut,
-                                               const uint8_t clustSizeCut) {
+                                               const uint16_t clustSizeCut) {
     if (modules.subdets()[lowerModuleIndex] == Barrel) {
       return runMiniDoubletDefaultAlgoBarrel(acc,
                                              modules,
@@ -697,7 +697,7 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE::lst {
                                   MiniDoubletsOccupancy mdsOccupancy,
                                   ObjectRangesConst ranges,
                                   const float ptCut,
-                                  const uint8_t clustSizeCut) const {
+                                  const uint16_t clustSizeCut) const {
       for (uint16_t lowerModuleIndex : cms::alpakatools::uniform_elements_y(acc, modules.nLowerModules())) {
         uint16_t upperModuleIndex = modules.partnerModuleIndices()[lowerModuleIndex];
         int nLowerHits = hitsRanges.hitRangesnLower()[lowerModuleIndex];
